requirements.txt
Pillow==10.1.0
ttkthemes==3.2.2
psd_tools==1.9.28